+++ src/components/HeroImage/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Button from 'react-bootstrap/Button';
-import './../assets/styles/styles.css';
+import './../../assets/styles/styles.css';
 
 function HeroImage(props) {
   return (
