+++ src/components/PortfolioPage/index.js
@@ -1,23 +1,23 @@
 import React from 'react';
-import HeroImage from './HeroImage';
-import HorizontalRule from './HorizontalRule';
-import Credit from './Credit';
+import HeroImage from '../HeroImage';
+import HorizontalRule from '../HorizontalRule';
+import Credit from '../Credit';
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
-import ImageCard from './../components/ImageCard';
-import passwordGenerator from './../assets/images/passwordGenerator.png';
-import codeQuiz from './../assets/images/codeQuiz.png';
-import workdayScheduler from './../assets/images/workdayScheduler.png';
-import weatherDashboard from './../assets/images/weatherDashboard.png';
-import goodReadme from './../assets/images/goodReadme.png';
-import teamProfile from './../assets/images/teamProfile.png';
-import noteTaker from './../assets/images/noteTaker.png';
-import employeeManagement from './../assets/images/employeeManagement.png';
-import fitnessTracker from './../assets/images/fitnessTracker.png';
-import employeeDirectory from './../assets/images/employeeDirectory.png';
-import petAdoption from './../assets/images/petAdoption.png';
-import blogSite from './../assets/images/blogSite.png';
-import planet from './../assets/images/planet.png';
+import ImageCard from '../ImageCard';
+import passwordGenerator from './../../assets/images/passwordGenerator.png';
+import codeQuiz from './../../assets/images/codeQuiz.png';
+import workdayScheduler from './../../assets/images/workdayScheduler.png';
+import weatherDashboard from './../../assets/images/weatherDashboard.png';
+import goodReadme from './../../assets/images/goodReadme.png';
+import teamProfile from './../../assets/images/teamProfile.png';
+import noteTaker from './../../assets/images/noteTaker.png';
+import employeeManagement from './../../assets/images/employeeManagement.png';
+import fitnessTracker from './../../assets/images/fitnessTracker.png';
+import employeeDirectory from './../../assets/images/employeeDirectory.png';
+import petAdoption from './../../assets/images/petAdoption.png';
+import blogSite from './../../assets/images/blogSite.png';
+import planet from './../../assets/images/planet.png';
 
 function PortfolioPage() {
   return (
